ai_attendance_system/requirements.txt
Flask
Werkzeug
face-recognition
opencv-python
numpy
Pillow
python-dotenv
scipy
